+++ routes/main.py
@@ -1,0 +1,8 @@
+from flask import Blueprint, render_template
+
+main_bp = Blueprint('main', __name__)
+
+
+@main_bp.route('/')
+def index():
+    return render_template('index.html')
