+++ templates/index.html
@@ -24,8 +24,8 @@
 
     <div class="hero-buttons">
       <!-- Get Started goes to signup, Login goes to login -->
-      <a href="{{ url_for('signup') }}" class="btn-primary">Get Started →</a>
-      <a href="{{ url_for('login') }}" class="btn-secondary">Login</a>
+      <a href="{{ url_for('auth.signup') }}" class="btn-primary">Get Started →</a>
+      <a href="{{ url_for('auth.login') }}" class="btn-secondary">Login</a>
     </div>
 
   </div>
